BST Q.2
Program in C language to search data using Binary Search Tree.
Inputs:
a. Data string as much as 10 pieces
b. Entered by user
c. The data sought is inputted by the user
Outputs:
a. Location of found data nodes
b. The message "not found" if the data is not found
Use modular programming and be user friendly